+++ src/mruby_integration/models/gamepad.cpp
@@ -68,6 +68,17 @@ mrb_Gamepad_name(mrb_state* mrb, mrb_value self) -> mrb_value
   return mrb_str_new_cstr(mrb, GetGamepadName(gamepad->index));
 }
 
+auto
+mrb_Gamepad_pressed(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  mrb_get_self(mrb, self, Gamepad, gamepad);
+
+  mrb_int button;
+  mrb_get_args(mrb, "i", &button);
+
+  return mrb_bool_value(IsGamepadButtonPressed(gamepad->index, button));
+}
+
 void
 append_models_Gamepad(mrb_state* mrb)
 {
@@ -81,6 +92,8 @@ append_models_Gamepad(mrb_state* mrb)
     mrb, Gamepad_class, "available?", mrb_Gamepad_available, MRB_ARGS_NONE());
   mrb_define_method(
     mrb, Gamepad_class, "name", mrb_Gamepad_name, MRB_ARGS_NONE());
+  mrb_define_method(
+    mrb, Gamepad_class, "pressed?", mrb_Gamepad_pressed, MRB_ARGS_REQ(1));
 
   load_ruby_models_gamepad(mrb);
 }